Split Lanczos files
